Move filter pipeline support functions/classes to new file
