+++ tiledb/sm/filter/test/unit_run_filter_pipeline.cc
@@ -79,6 +79,7 @@
 #include "add_1_including_metadata_filter.h"
 #include "add_1_out_of_place_filter.h"
 #include "add_n_in_place_filter.h"
+#include "filter_test_support.h"
 #include "filtered_tile_checker.h"
 #include "pseudo_checksum_filter.h"
 #include "tile_data_generator.h"
@@ -91,171 +92,6 @@
 #include "tiledb/sm/tile/tile.h"
 
 using namespace tiledb::sm;
-
-// A dummy `Stats` instance.
-static tiledb::sm::stats::Stats dummy_stats("test");
-
-class SimpleTestData {};
-
-/**
- * Original variable length test from the pipeline tests.
- *
- * For this test target size is 10 cells per chunk. Below is a list of value
- * cell lengths, the cell they are added to, and the rational.
- *
- * target = 8 cells
- * min = 4 cells
- * max = 12 cells
- *
- * | # Cells | Previous/New # Cell in Chunk | Notes                            |
- * |:-------:|:--------|:------------------------------------------------------|
- * |  4      |  0 / 4  | chunk 0: initial chunk                                |
- * |  10     |  4 / 14 | chunk 0: new > max, prev. <= min (next new)           |
- * |  6      |  0 / 6  | chunk 1: new <= target                                |
- * |  11     |  6 / 11 | chunk 2: target < new <= max, prev. > min  (next new) |
- * |  7      |  0 / 7  | chunk 3: new <= target                                |
- * |  9      |  7 / 16 | chunk 4: new > max, prev. > min (this new)            |
- * |  1      |  9 / 10 | chunk 4: new <= target                                |
- * |  10     | 10 / 20 | chunk 5: new > max, prev. > min (this new)            |
- * |  20     |  0 / 20 | chunk 6: new > max, prev. < min (next new)            |
- * |  2      |  0 / 2  | chunk 7: new <= target                                |
- * |  2      |  2 / 4  | chunk 7: new <= target                                |
- * |  2      |  4 / 6  | chunk 7: new <= target                                |
- * |  2      |  6 / 8  | chunk 7: new <= target                                |
- * |  2      |  8 / 10 | chunk 7: new <= target                                |
- * |  12     | 10 / 24 | chunk 8: new > max, prev. > min (this new)            |
- *
- */
-class SimpleVariableTestData {
- public:
-  SimpleVariableTestData()
-      : target_ncells_per_chunk_{10}
-      , elements_per_chunk_{14, 6, 11, 7, 10, 10, 20, 10, 12}
-      , tile_data_generator_{
-            {4, 10, 6, 11, 7, 9, 1, 10, 20, 2, 2, 2, 2, 2, 12}} {
-    WriterTile::set_max_tile_chunk_size(
-        target_ncells_per_chunk_ * sizeof(uint64_t));
-  }
-  ~SimpleVariableTestData() {
-    WriterTile::set_max_tile_chunk_size(constants::max_tile_chunk_size);
-  }
-
-  /** Returns the number elements (cells) stored in each chunk. */
-  const std::vector<uint64_t>& elements_per_chunk() const {
-    return elements_per_chunk_;
-  }
-
-  const TileDataGenerator& tile_data_generator() const {
-    return tile_data_generator_;
-  }
-
- private:
-  uint64_t target_ncells_per_chunk_{};
-  std::vector<uint64_t> elements_per_chunk_{};
-  IncrementTileDataGenerator<uint64_t, Datatype::UINT64> tile_data_generator_;
-};
-
-/**
- * Checks the following:
- *
- * 1. Pipeline runs forward without error.
- * 2. Filtered buffer data is as expected.
- * 3. Pipeline runs backward without error.
- * 4. Result from roundtrip matches the original data.
- */
-void check_run_pipeline_full(
-    Config& config,
-    ThreadPool& tp,
-    shared_ptr<WriterTile>& tile,
-    std::optional<shared_ptr<WriterTile>>& offsets_tile,
-    FilterPipeline& pipeline,
-    const TileDataGenerator* test_data,
-    const FilteredTileChecker& filtered_buffer_checker,
-    shared_ptr<MemoryTracker> memory_tracker) {
-  // Run the pipeline forward.
-  CHECK(pipeline
-            .run_forward(
-                &dummy_stats,
-                tile.get(),
-                offsets_tile.has_value() ? offsets_tile.value().get() : nullptr,
-                &tp)
-            .ok());
-
-  // Check the original unfiltered data was removed.
-  CHECK(tile->size() == 0);
-
-  // Check the filtered buffer has the expected data.
-  auto filtered_buffer = tile->filtered_buffer();
-  filtered_buffer_checker.check(filtered_buffer);
-
-  // Run the data in reverse.
-  auto unfiltered_tile =
-      test_data->create_filtered_buffer_tile(filtered_buffer, memory_tracker);
-  ChunkData chunk_data;
-  unfiltered_tile.load_chunk_data(chunk_data);
-  CHECK(pipeline
-            .run_reverse(
-                &dummy_stats,
-                &unfiltered_tile,
-                nullptr,
-                chunk_data,
-                0,
-                chunk_data.filtered_chunks_.size(),
-                tp.concurrency_level(),
-                config)
-            .ok());
-
-  // Check the original data is reverted.
-  test_data->check_tile_data(unfiltered_tile);
-}
-
-/**
- * Checks the following:
- *
- * 1. Pipeline runs forward without error.
- * 2. Pipeline runs backward without error.
- * 3. Result from roundtrip matches the original data.
- */
-void check_run_pipeline_roundtrip(
-    Config& config,
-    ThreadPool& tp,
-    shared_ptr<WriterTile> tile,
-    std::optional<shared_ptr<WriterTile>>& offsets_tile,
-    FilterPipeline& pipeline,
-    const TileDataGenerator* test_data,
-    shared_ptr<MemoryTracker> memory_tracker) {
-  // Run the pipeline forward.
-  CHECK(pipeline
-            .run_forward(
-                &dummy_stats,
-                tile.get(),
-                offsets_tile.has_value() ? offsets_tile.value().get() : nullptr,
-                &tp)
-            .ok());
-
-  // Check the original unfiltered data was removed.
-  CHECK(tile->size() == 0);
-
-  // Run the data in reverse.
-  auto unfiltered_tile = test_data->create_filtered_buffer_tile(
-      tile->filtered_buffer(), memory_tracker);
-  ChunkData chunk_data;
-  unfiltered_tile.load_chunk_data(chunk_data);
-  CHECK(pipeline
-            .run_reverse(
-                &dummy_stats,
-                &unfiltered_tile,
-                nullptr,
-                chunk_data,
-                0,
-                chunk_data.filtered_chunks_.size(),
-                tp.concurrency_level(),
-                config)
-            .ok());
-
-  // Check the original data is reverted.
-  test_data->check_tile_data(unfiltered_tile);
-}
 
 TEST_CASE("Filter: Test empty pipeline", "[filter][empty-pipeline]") {
   // Create TileDB needed for running pipeline.
